Backend_API/requirements.txt
anyio==3.6.1
fastapi==0.85.0
uvicorn==0.18.3
numpy==1.23.3
pandas==1.5.0
pydantic==1.10.2
google-big-query==3.9.0
pandas-gbq==0.19.1